refactor: Position, RawPosition, Route, RawRoute 패키지 분리
