+++ src/main/java/janggi/domain/piece/General.java
@@ -1,7 +1,9 @@
 package janggi.domain.piece;
 
-import janggi.domain.RawRoute;
 import janggi.domain.Team;
+import janggi.domain.position.Position;
+import janggi.domain.position.RawPosition;
+import janggi.domain.position.RawRoute;
 import java.util.List;
 import java.util.Set;
