+++ src/main/java/janggi/domain/position/RawPosition.java
@@ -1,4 +1,4 @@
-package janggi.domain.piece;
+package janggi.domain.position;
 
 public record RawPosition(int x, int y) {
