+++ src/test/java/janggi/domain/piece/HorseTest.java
@@ -3,7 +3,8 @@ package janggi.domain.piece;
 import static janggi.domain.Team.RED;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import janggi.domain.Route;
+import janggi.domain.position.Position;
+import janggi.domain.position.Route;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
